+++ cloud/DAL/RPMWeb.Data.Common/Data.cs
@@ -564,7 +564,7 @@ namespace RPMWeb.Data.Common
         public string DOB { get; set; }
 
         public string Gender { get; set; }
-        public int Height { get; set; }
+        public float Height { get; set; }
         public int Weight { get; set; }
         public string Email { get; set; }
         public string PhoneNo { get; set; }
